Moving some code around to be a bit less painful.
